Add some documentation for CapturingTessellator
