+++ src/main/java/com/gtnewhorizons/angelica/glsm/TessellatorManager.java
@@ -1,12 +1,9 @@
 package com.gtnewhorizons.angelica.glsm;
 
 import com.gtnewhorizons.angelica.client.renderer.CapturingTessellator;
-import com.gtnewhorizons.angelica.compat.mojang.VertexFormat;
 import com.gtnewhorizons.angelica.compat.nd.Quad;
 import net.minecraft.client.renderer.Tessellator;
-import org.lwjgl.BufferUtils;
 
-import java.nio.ByteBuffer;
 import java.util.List;
 
 @SuppressWarnings("unused")
@@ -32,6 +29,7 @@ public class TessellatorManager {
     }
 
     public static void startCapturing(CapturingTessellator tessellator) {
+        if(capturingTessellator != null) throw new IllegalStateException("Tried to start capturing when already capturing!");
         capturingTessellator = tessellator;
     }
     public static List<Quad> stopCapturing() {
@@ -40,30 +38,6 @@ public class TessellatorManager {
         final List<Quad> quads = capturingTessellator.getQuads();
         capturingTessellator = null;
         return quads;
-    }
-
-    public static ByteBuffer quadsToBuffer(List<Quad> quads, VertexFormat format) {
-        final ByteBuffer byteBuffer = BufferUtils.createByteBuffer(format.getVertexSize() * quads.size() * 4);
-
-        // noinspection ForLoopReplaceableByForEach
-        for (int i = 0, quadsSize = quads.size(); i < quadsSize; i++) {
-            final Quad quad = quads.get(i);
-            for (int idx = 0; idx < 4; idx++) {
-                // Position
-                byteBuffer.putFloat(quad.getX(idx));
-                byteBuffer.putFloat(quad.getY(idx));
-                byteBuffer.putFloat(quad.getZ(idx));
-
-                // Texture
-                byteBuffer.putFloat(quad.getTexU(idx));
-                byteBuffer.putFloat(quad.getTexV(idx));
-
-                // Normals
-                byteBuffer.putInt(quad.getNormal(idx));
-            }
-        }
-        byteBuffer.rewind();
-        return byteBuffer;
     }
 
     static {
